+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.38.16
 
+* `Core`:
+    * `TelegramAPIUrlsKeeper` now have two new things: properties `webAppDataSecretKey` and fun `checkWebAppLink`
+
 ## 0.38.15
 
 * `Common`: